refactor: file name, 不要なコメントアウトの削除
